Bug found by Thomas Rapp using the stripetest driver for raidlib - return code not covered
Fix is to check the return code and return an ERROR (-1) if the stripe input file can't be opened for some reason - e.g. does not exist.  This is not a user error as we don't really want to simply segfault if they spell the filename wrong or just leave off the filename extension, etc., we'd instead prefer this to exit gracefully or bubble up the error to the test driver or application to handle.