Checking all file open system calls and either adding error recovery or assert
To handle a bug (user error) without a segfault using a bad file pointer, error handling and/or assert is used in case a bad argument is provided.  This could be improved with more sophisticated user interaction, but for now, this at lease either recovers or positively asserts if there is no obvious recovery.